+++ tests/ModifiedMatcherTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace SmartonDev\HttpCache\Tests2;
+namespace SmartonDev\HttpCache\Tests;
 
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
@@ -69,30 +69,31 @@ class ModifiedMatcherTest extends TestCase
         $this->assertSame($expectedIfUnmodifiedSinceAsTimestamp, $matcher->getIfUnmodifiedSinceHeaderAsTimestamp());
     }
 
-    public function testMatches(): void
+    public function testMatchesIfModifiedSince(): void
     {
         $matcher = (new ModifiedMatcher())->headers([
             'If-Modified-Since' => 'Tue, 15 Nov 1994 12:45:26 GMT',
+        ]);
+
+        $dtBefore = new \DateTime('Mon, 14 Nov 1994 12:45:26 GMT');
+        $dtAfter = new \DateTime('Wed, 16 Nov 1994 12:45:26 GMT');
+        $dtEq = new \DateTime('Tue, 15 Nov 1994 12:45:26 GMT');
+        $this->assertTrue($matcher->matches($dtEq)->matchesModifiedAt());
+        $this->assertFalse($matcher->matches($dtBefore)->matchesModifiedAt());
+        $this->assertTrue($matcher->matches($dtAfter)->isModifiedSince());
+    }
+
+    public function testMatchesIfUnmodifiedSince(): void
+    {
+        $matcher = (new ModifiedMatcher())->headers([
             'If-Unmodified-Since' => 'Tue, 15 Nov 1994 12:45:26 GMT',
         ]);
 
         $dtBefore = new \DateTime('Mon, 14 Nov 1994 12:45:26 GMT');
         $dtAfter = new \DateTime('Wed, 16 Nov 1994 12:45:26 GMT');
         $dtEq = new \DateTime('Tue, 15 Nov 1994 12:45:26 GMT');
-        $this->assertTrue($matcher->matches($dtEq)->isEqualsModifiedSince());
-        $this->assertFalse($matcher->matches($dtBefore)->isEqualsModifiedSince());
-        $this->assertTrue($matcher->matches($dtBefore)->isBeforeModifiedSince());
-        $this->assertFalse($matcher->matches($dtAfter)->isBeforeModifiedSince());
-        $this->assertFalse($matcher->matches($dtAfter)->isEqualsModifiedSince());
-        $this->assertFalse($matcher->matches($dtBefore)->isAfterModifiedSince());
-        $this->assertTrue($matcher->matches($dtAfter)->isAfterModifiedSince());
-
-        $this->assertTrue($matcher->matches($dtEq)->isEqualsUnmodifiedSince());
-        $this->assertFalse($matcher->matches($dtBefore)->isEqualsUnmodifiedSince());
-        $this->assertTrue($matcher->matches($dtBefore)->isBeforeUnmodifiedSince());
-        $this->assertFalse($matcher->matches($dtAfter)->isBeforeUnmodifiedSince());
-        $this->assertFalse($matcher->matches($dtAfter)->isEqualsUnmodifiedSince());
-        $this->assertFalse($matcher->matches($dtBefore)->isAfterUnmodifiedSince());
-        $this->assertTrue($matcher->matches($dtAfter)->isAfterUnmodifiedSince());
+        $this->assertTrue($matcher->matches($dtBefore)->isUnmodifiedSince());
+        $this->assertFalse($matcher->matches($dtAfter)->isUnmodifiedSince());
+        $this->assertFalse($matcher->matches($dtEq)->isModifiedSince());
     }
 }